+++ packages/larva/lib/server.js
@@ -17,7 +17,6 @@ const getPatternData = require( './utils/getPatternData' );
 const getSubDirectoryNames = require( './utils/getSubDirectoryNames' );
 
 const app = express();
-const port = process.env.NODE_PORT || 3000;
 
 const appConfiguration = getAppConfiguration( 'patterns' );
 const twigPaths = getPatternPathsToLoad( appConfiguration );
@@ -189,7 +188,6 @@ app.get( '/:source/:type/:name/:variant?', function (req, res) {
 
 	// Support query parameters for conditionally loading stylesheets and scripts
 	req.params[ 'query' ] = req.query;
-
 	req.params[ 'data' ] = getPatternData( patternsPath, req.params );
 	req.params[ 'pattern_nav' ] = patterns;
 
@@ -200,6 +198,4 @@ app.get( '/:source/:type/:name/:variant?', function (req, res) {
 	res.end( twing.render( 'pattern.html', req.params ) );
 });
 
-app.listen(port, () => {
-	console.log( 'Larva server is listening on port ' + port );
-});
+module.exports = app;